AdminController finished

+++ src/main/java/com/bag/hotelmanagementsystem/model/UserModel.java
@@ -33,7 +33,7 @@ public class UserModel
     private Long tcNumber;
 
     @Column(name = "telephonenumber",nullable=false)
-    private int telephoneNumber;
+    private Long telephoneNumber;
 
     @Column(name = "isAdmin",nullable=true)
     private boolean isAdmin;
